Tidy up app directory structure
